First submission of Missile Simulation
This simulation setup is meant to allow for trying out ideas with mock
missiles and targets. Current developments have built an Equations of
Motions class which essentially represents kinematic changes in an
object’s state, such as position, velocity, attitude, and angular
rates. Its formulation is based on standard dynamical system equations
(Velocity and Acceleration in a rotating frame, quaternion derivatives,
Euler Equations for rigid bodies).

An object’s position and velocity are updated in the Earth Centered
Earth Fixed Frame to account for Earth’s curvature and rotation via
WGS84 Models.

There’s a lot of work left to do in designing a missile, targets,
sensors, and subsystems for each.